scripts/requirements.txt
boto3>=1.34
psycopg2-binary>=2.9
faker>=24.0
pandas>=2.2
great-expectations>=0.18,<1.0
requests>=2.31
python-dotenv>=1.0
pyyaml>=6.0
tabulate>=0.9
